% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R, R/irods.R
\name{sodar_assay_irods_path}
\alias{sodar_assay_irods_path}
\alias{sodar_study_irods_path}
\title{Retrieve the iRODS path of an assay or study}
\usage{
sodar_assay_irods_path(
  project_uuid,
  study_uuid = NULL,
  assay_uuid = NULL,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_study_irods_path(
  project_uuid,
  study_uuid = NULL,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_assay_irods_path(
  project_uuid,
  study_uuid = NULL,
  assay_uuid = NULL,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_study_irods_path(
  project_uuid,
  study_uuid = NULL,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project.}

\item{study_uuid}{A character string containing the UUID of the study.}

\item{assay_uuid}{A character string containing the UUID of the assay.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
A character string containing the iRODS path of the assay.

A character string containing the iRODS path of the assay.
}
\description{
This functions retrieves the iRODS path of an assay or study.

This functions retrieves the iRODS path of an assay or study.
}
\details{
It is not necessary to provide the assay UUID if there is only one assay in the project.

It is not necessary to provide the assay UUID if there is only one assay in the project.
}
