% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{sodar_category_list}
\alias{sodar_category_list}
\title{Returns a list of all projects and categories belonging to a category}
\usage{
sodar_category_list(
  category_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{category_uuid}{A character string containing the UUID of the category.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
A color data frame (colorDF) containing the SODAR project list.
}
\description{
This function returns a list of all projects and categories belonging to a given category.
}
