% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{sodar_config}
\alias{sodar_config}
\title{Set the SODAR configuration.}
\usage{
sodar_config(sodar_api_token, sodar_url, set_options = TRUE)
}
\arguments{
\item{sodar_api_token}{A character string containing the SODAR API token.}

\item{sodar_url}{A character string containing the SODAR URL.}

\item{set_options}{A logical value indicating whether to set the global options.}
}
\value{
An invisible list containing the SODAR configuration.
}
\description{
This function sets the SODAR configuration.
}
