% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R, R/irods.R
\name{sodar_create_irods}
\alias{sodar_create_irods}
\title{Create an iRODS collection for a project}
\usage{
sodar_create_irods(
  project_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_create_irods(
  project_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\description{
This function creates an iRODS collection for a project.

This function creates an iRODS collection for a project.
}
