% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods.R
\name{sodar_delete_request}
\alias{sodar_delete_request}
\alias{sodar_request_list}
\alias{sodar_request_cancel}
\alias{sodar_request_reject}
\alias{sodar_request_accept}
\title{Create and process iRODS data requests}
\usage{
sodar_delete_request(
  project_uuid,
  path,
  description = NULL,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_list(
  project_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_cancel(
  request_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_reject(
  request_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_accept(
  request_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project.}

\item{path}{A character string containing the path to the object or collection.}

\item{description}{A character string containing the description of the request.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}

\item{request_uuid}{A character string containing the UUID of the request
(returned by sodar_delete and sodar_request_list).}
}
\description{
Create an iRODS delete request for an object or collection
}
