% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{sodar_file_exists}
\alias{sodar_file_exists}
\title{Check if a file exists in iRODS using md5 checksum}
\usage{
sodar_file_exists(
  md5,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{md5}{A character vector containing md5 checksums.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
A named logical vector indicating whether the file exists.
}
\description{
This function checks if a file exists in iRODS using the md5 checksum.
}
