% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{sodar_file_list}
\alias{sodar_file_list}
\title{List files present in a project sample data repository}
\usage{
sodar_file_list(
  project_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\description{
List files present in a project sample data repository
}
