% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{sodar_investigation_retrieve}
\alias{sodar_investigation_retrieve}
\title{Retrieve details of an investigation from a project}
\usage{
sodar_investigation_retrieve(
  project_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
A list with the details of the investigation.
}
\description{
This function retrieves details of an investigation from a project. This
can be used to retrieve assay UUIDs for landing zone operations.
}
