% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods.R
\name{sodar_irods_configuration}
\alias{sodar_irods_configuration}
\title{Retrieve the iRODS configuration}
\usage{
sodar_irods_configuration(
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
A list with the iRODS configuration.
}
\description{
Retrieve the iRODS configuration
}
