% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{sodar_lz_create}
\alias{sodar_lz_create}
\title{Create a landing zone for a project}
\usage{
sodar_lz_create(
  project_uuid,
  create_colls = TRUE,
  restrict_colls = TRUE,
  title = "",
  assay_uuid = NULL,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project.}

\item{create_colls}{A logical value indicating whether to create collections for the landing zone.}

\item{restrict_colls}{A logical value indicating whether to restrict collections for the landing zone.}

\item{title}{A character string containing the title suffix of the landing zone.}

\item{assay_uuid}{A character string containing the UUID of the assay to associate with
the landing zone. If NULL, the first assay found will be used.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
A list with the details of the landing zone created.
}
\description{
This function creates a landing zone for a project.
}
\seealso{
sodar_lz_retrieve sodar_lz_list
}
