% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz.R
\name{sodar_lz_list}
\alias{sodar_lz_list}
\title{List landing zones for a project}
\usage{
sodar_lz_list(
  project_uuid,
  verbose = getOption("verbose", default = FALSE),
  config = NULL
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}
}
\value{
A colorDF data frame with the details of the landing zones.
Several columns are hidden for better readability, but can still be
accessed.
}
\description{
This function lists landing zones for a project.
}
\seealso{
sodar_lz_retrieve sodar_lz_create
}
