% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{sodar_lz_retrieve}
\alias{sodar_lz_retrieve}
\alias{sodar_lz_status}
\alias{sodar_lz_delete}
\alias{sodar_lz_move}
\alias{sodar_lz_validate}
\title{Handling existing landing zones}
\usage{
sodar_lz_retrieve(
  lz_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_lz_status(
  lz_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE)
)

sodar_lz_delete(
  lz_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_lz_move(
  lz_uuid,
  move = TRUE,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_lz_validate(
  lz_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{lz_uuid}{A character string containing the UUID of the landing zone.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}

\item{move}{A logical value indicating whether to move the landing zone
after validation.}
}
\value{
A list with the details of the landing zone.
}
\description{
This functions retrieve information about a landing zone for a project,
delete existing landing zones etc.
}
