% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{sodar_project_create}
\alias{sodar_project_create}
\title{Create a new project in SODAR}
\usage{
sodar_project_create(
  title,
  parent,
  type = "PROJECT",
  description = NULL,
  readme = NULL,
  public_guest_access = FALSE,
  owner = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE,
  config = NULL
)
}
\arguments{
\item{title}{A character string containing the title of the project or category; mandatory.}

\item{parent}{A character string containing the UUID of the parent project or category; mandatory.}

\item{type}{A character string indicating whether it is a project or a category.}

\item{description}{A character string containing the description of the project or category (optional).}

\item{readme}{A character string containing the README of the project or category (optional).}

\item{public_guest_access}{A logical value indicating whether the project or category has public guest access.}

\item{owner}{A character string containing the UUID of the owner of the project or category. If NULL, the
current user (indicated by the API token) will be set as owner.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}
}
\value{
A list with the details of the project created.
}
\description{
This function creates a new project or category in SODAR.
}
