% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{sodar_project_list}
\alias{sodar_project_list}
\title{Get the SODAR project list table.}
\usage{
sodar_project_list(
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
A color data frame (colorDF) containing the SODAR project list.
}
\description{
This function gets the SODAR project list table.
}
\details{
The function returns a data frame flavored by the colorDF package. This allows
to hide multiline columns such as "readme", "roles" or "description". The columns
are still present in the data frame and can be accessed as usual, but
they are not shown in the console.
}
