% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{sodar_project_retrieve}
\alias{sodar_project_retrieve}
\title{Retrieve details of a project or category}
\usage{
sodar_project_retrieve(
  project_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project or category.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
A list with the details of the project or category.
}
\description{
This function retrieves details of a project or category.
}
