% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{sodar_project_update}
\alias{sodar_project_update}
\title{Update a project or category in SODAR}
\usage{
sodar_project_update(
  project_uuid,
  title = NULL,
  parent = NULL,
  description = NULL,
  readme = NULL,
  public_guest_access = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE,
  config = NULL
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project or category.}

\item{title}{A character string containing the title of the project or category; mandatory.}

\item{parent}{A character string containing the UUID of the parent project or category; mandatory.}

\item{description}{A character string containing the description of the project or category (optional).}

\item{readme}{A character string containing the README of the project or category (optional).}

\item{public_guest_access}{A logical value indicating whether the project or category has public guest access.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}
}
\value{
A list with the details of the project updated.
}
\description{
This function updates a project or category in SODAR.
}
