% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{sodar_sheet_export_zip}
\alias{sodar_sheet_export_zip}
\alias{sodar_sheet_export_json}
\title{Export an ISAtab sample sheet from a project}
\usage{
sodar_sheet_export_zip(
  project_uuid,
  file,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_sheet_export_json(
  project_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project to export.}

\item{file}{A character string containing the name of the file to be exported.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
\itemize{
\item sodar_sheet_export_zip: The name of the file exported.
\item sodar_sheet_export_json: A list with the exported isatab structure.
}
}
\description{
This function exports an ISAtab sample sheet from a project.
}
