% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{sodar_user_role_create}
\alias{sodar_user_role_create}
\alias{sodar_user_role_update}
\title{Create or modify a new user in a project}
\usage{
sodar_user_role_create(
  project_uuid,
  user_uuid,
  role,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_user_role_update(
  project_uuid,
  user_uuid,
  role,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{project_uuid}{A character string containing the UUID of the project.}

\item{user_uuid}{A character string containing the UUID of the user.
This must be an existing user in the SODAR instance.}

\item{role}{A character string containing the role of the user in the project.}

\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\value{
A list with the details of the user created.
}
\description{
These functions create  a new user in a project or modify the role of an existing user.
}
\seealso{
sodar_user_list
}
