% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{sodar_whoami}
\alias{sodar_whoami}
\title{Returns information about the user making the request}
\usage{
sodar_whoami(
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)
}
\arguments{
\item{config}{A list containing the SODAR configuration. If NULL, it tries to get it from the global options.}

\item{verbose}{A logical value indicating whether to make verbose requests.}

\item{return_raw}{A logical value indicating whether to return the raw response.}
}
\description{
Returns information about the user making the request.
}
